\name{airline_GP_prediction-package}
\alias{airline_GP_prediction-package}
\alias{airline_GP_prediction}
\docType{package}
\title{
\packageTitle{airline_GP_prediction}
}
\description{
\packageDescription{airline_GP_prediction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{airline_GP_prediction}
\packageIndices{airline_GP_prediction}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{airline_GP_prediction}

Maintainer: \packageMaintainer{airline_GP_prediction}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
